requirements.txt
# News API
newsapi-python==0.2.7
requests==2.31.0

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Sentiment Analysis
vaderSentiment==3.3.2
textblob==0.17.1

# Database
pymongo==4.6.1

# Visualization
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
wordcloud==1.9.3

# Utilities
python-dotenv==1.0.0
pytz==2023.3

# Logging
loguru==0.7.2



# Optional: Advanced NLP
# transformers==4.35.2
# torch==2.1.2